+++ src/028-number-spiral-diagonals/c/src/solve.c
@@ -1,6 +1,6 @@
 #include "solve.h"
 
-uint64_t solve(size_t max)
+uint64_t solve(uint64_t max)
 {
     uint64_t sum = 1;
     uint64_t cur = 1;
