+++ src/028-number-spiral-diagonals/c/src/solve.h
@@ -1,5 +1,4 @@
-#include <stddef.h>
 #include <stdint.h>
 #pragma once
 
-uint64_t solve(size_t max);
+uint64_t solve(uint64_t max);
