+++ src/036-doublebase-palindromes/c/src/solve.h
@@ -3,5 +3,5 @@
 #include <stdbool.h>
 #pragma once
 
-uint32_t solve(size_t max);
-bool is_palindrome(uint32_t num, size_t base);
+uint32_t solve(uint32_t max);
+bool is_palindrome(uint32_t num, uint32_t base);
